Fixing multi-byte controls and a big refactoring to remove the raw layer
Fixing the multi-byte controls loading from SysEx dumps it transpired that
the raw layer as it is is not flaxible enough. Value from MIDI scaling must
take into account the u16 range of the target control.

Raw layer removed completely, replaced with a ProgramsDump paged structure
for containing programs and loading them from/to the active controller.