Added program name editing in the UI
This is unfortunately a bit of a hack. For now the name change
is signalled through a special "name_change" signal in the edit
buffer controller.

The program name is read from/written to the raw layer directly,
so there's no signalling channel to hook up.

NOTE: If the `Raw` is brought back, there can be legit
"controller -> raw" and "raw -> controller" signalling for this!