Added ArcSwap to properly propagate edit buffer and dump from modules to code
Edit buffer and dump `Arc`s may already have been cloned by the time
are returned as a part of `InitializedInterface`, so the contents
of the `Arc` cannot be just moved out of the intialized interface struct.

Access throughout the main code can't be `Arc<Mutex<EditBuffer>>`, but
should instead be `Arc<Arc<Mutex<EditBuffer>>>`. Instead we add `ArcSwap`
to make the internal `Arc` change atomic.