A new implementation to get sysex message length to work around broken drivers
PocketPOD driver with WinRT backend returns the full sysex buffer (136 bytes)
as opposed to the 0xf0 .. 0xf7 like it is done with all other backends.
To work around the broken driver/backend, the code now checka the incoming
sysex message for 0xf7 manually and -for good measure- for any other
0x80-containing bytes that shouldn't be there in the first place (cancelling
the message).