Implemented generic dynamic toggles positioning
Implemented generic code that takes a list of toggles and a reference to
the grid object to be used and sets up the toggles in the grid according
to the provided position data. Toggles that change position based on some
control (moving) get a "<"/">" button and wiring to change position
according to the UI action or control change.

This replaces hard-coded volume pedal toggle positionining in POD 2.0.