Close MIDI ports when opening the settings dialog
The main reason to do this is to make the "Autodetect" and "Test" buttons
work in the windows (WinMM) build, where a MIDI port may only be open once.
If the app hold the ports open (after a successful auto-detect on start-up,
for example, then the "Test" button won't work as we won't be able to open
the port for testing...

The logic that closes the ports when the settings dialog opens and reopens
them when the settings dialog closes works around this nonsence.

This will also work better when the POD connects to a loopback device
like a MIDI through port in Linux, so that the app doesn't answer its own
MIDI queries during autodetection.

re #10