Quiet time on MIDI in before closing PocketPOD MIDI channel when using WinMM
Implemented logic for PocketPOD when using WinMM back-end to wait for 1 second
of quiet on the MIDI in port before closing, otherwise the app hangs (see #12).

This also changes a Quit UI event into a two-step ShutDown (which triggers MIDI
ports closing) and Quit (once the MIDI ports are closed).