Store trait changed so that Event also carries value
There was a problem with the way Store event receivers first receive the
event and then go reading the value from the controller. At this point,
the value might've changed already (as in the case with the PODxt, when
turning the "effects" knob chaninges `effect_select` and triggers an
edit buffer read from the device, so by the time `effect_select` event
is read, the controller value may have already been overwritten from
edit buffer dump with a 0).

Store/Event logic rewritten so that the Event also carries a value,
so that there's no need to reach to the controller for the value.
May have inadvertently fixed some other similar subtle bugs with this.