Don't show an active "Tuner" button on UI start-up
This has been a minor annoyance since the beginning. The "tuner" button
was always shown as selected before the UI sets the "hidden" button. Now,
we don't show "Manual" or "Tuner" on startup, but also unset "active"
for either of them so that neither is selected when they do show.