Filter out unsupported characters in program name entry
Line6Edit allows non-ASCII and non-printable ASCII characters inside the
editor, but replaces them with space `0x20` or underscore `0x5f`. We'll
filter them out in the entry widget for now.

Also added trimmed "cur != new" check in the "controller -> gui" update
so that trailing space is allowed in the Entry while it is stripped
in the dump data (to allow backspace action when resulting code contains
a trailing white-space). re #28