POD 2.0 family: fix double handling of MIDI IN messages
Handling of MIDI IN messages handling from POD 2.0 handler was
duplicated in the generic handlier, which POD 2.0 handler also
calls. As a result, all messages such as `ProgramPatchDump`
and `ProgramPatchDumpRequest` were handled twice. As a result
of the latter, for example, the replies to the program
dumps were sent in double.

This affects all POD 2.0-family devices: POD 2.0, POD Pro,
POD 1 and also PocketPOD (also uses `Pod2Handler`)