POD 2.0 family: Fix possible crashes after an out-of-range CC 19 "effect_select:raw"
There is are valirous triggers (namely, "delay" and "effect_tweak") defined
in `mod-pod2/src/wiring.rs` that depend on a correct value of "effect_select:raw"
and assumed `effect_entry_for_value()` always returned a valid entry.
This may have been try once, but now it is an `Option<(...)>`. This leads to
crashes if effect select control sent an out-of-range value. And either the
delay is toggled in the UI or an effect tweak control sends a message.
This is likely the cause of #39.

Fixes #39