USB: delay opening USB device until enumeration in one of the USB lookup functions
Refactor USB code not to open USB devices as soon as they
register, but when needed for one of the lookup functions
(usb_list_devices, usb_device_for_name, usb_device_for_address).
This way, when a device open fails, it can be retried at
a later time, when USB enumeration happens again.

This is specifically useful in Linux when new devices are
owned by `root:root` and need a chmod to be usable by a
non-root user.