Add "no reply retry" count for autodetect/test
Seems like it takes a few messages to get the PODxt going. The test
and autodetect logic send a UDI message and expect a reply. When one
doesn't come, this means that no device answered, no device found.

With this change, the autodetect and test logic is changes to try
again if there was no reply (and no error) as many times as the
`MidiIn` requests. We set no reply retry count for MIDI devices to 0
and for USB devices to 3. Well-behaving USB device (PocketPOD) answer
right away, but this retry seems enough to get the PODxt talking.

PS. Maybe we just need to give PODxt more time after a reset &
configure to start talking, but let's try this first.