USB: Fix use-after-free bug around `Transfer::callback_inner`
Dropping a box while also holding a lock that is part of the
object in the box resulted in a use-after-free bug that reliably
manifested only on Apple Silicon and completely crashed the
app in all kinds of random functions (stack corruption of someone
else's stack?).

`RUSTFLAGS="-Z sanitizer=address"` in rust nightly is a savior!

fixes #66